+++ src/css/css_color_grammar_x3.cpp
@@ -20,7 +20,8 @@
  *
  *****************************************************************************/
 
-#include <mapnik/css_color_grammar_x3_def.hpp>
+#include <mapnik/css/css_grammar_x3.hpp>
+#include <mapnik/css/css_color_grammar_x3_def.hpp>
 #if BOOST_VERSION < 107000
 #include <mapnik/image_filter_types.hpp>
 #endif
@@ -29,8 +30,10 @@ namespace mapnik { namespace css_color_grammar {
 namespace x3 = boost::spirit::x3;
 using iterator_type = std::string::const_iterator;
 using context_type = x3::phrase_parse_context<x3::ascii::space_type>::type;
+using context_skipper_type = x3::phrase_parse_context<css_grammar::css_skipper_type>::type;
 
 BOOST_SPIRIT_INSTANTIATE(css_color_grammar_type, iterator_type, context_type);
+BOOST_SPIRIT_INSTANTIATE(css_color_grammar_type, iterator_type, context_skipper_type);
 
 #if BOOST_VERSION < 107000
 template bool parse_rule<iterator_type, context_type, mapnik::filter::color_to_alpha>
